// src/app/models/course-management/coursepublicWorkerData.ts
/**
 * Generated by orval v6.23.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface CoursepublicWorkerData {
    /** สถานะการใช้งาน */
    activeFlag?: boolean;
    /** FK (coursepublic_main) */
    coursepublicId?: number;
    /** PK */
    coursepublicWorkerId?: number;
    /** custom อีเมล */
    email?: string;
    /** อีเมล */
    externalEmail?: string;
    /** ชื่อบุคลากร (ภาษาอังกฤษ) */
    externalNameEn?: string;
    /** ชื่อบุคลากร (ภาษาไทย) */
    externalNameTh?: string;
    /** ชื่อบุคลากร (ส่วนงาน) */
    externalDepName?: string;
    first?: number;
    /** custom ชื่อบุคลากร (ภาษาอังกฤษ) */
    fullnameEn?: string;
    /** custom ชื่อบุคลากร (ภาษาไทย) */
    fullnameTh?: string;
    mode?: string;
    /** ตำแหน่งคณะทำงาน LOOKUP (30038000)  */
    positionType?: number;
    /** ตำแหน่งคณะทำงาน (ภาษาอังกฤษ)  */
    positionTypeEn?: string;
    /** ตำแหน่งคณะทำงาน (ภาษาไทย)  */
    positionTypeTh?: string;
    rowNum?: number;
    size?: number;
    /** FK (mas_personal) */
    workerId?: number;
    /** บุคลากรภายนอก */
    workerType?: boolean;

    depNameThLevel1?: string;
    depNameEnLevel1?: string;
    depNameThLevel2?: string;
    depNameEnLevel2?: string;

}