// src/app/models/master/masExternalPersonalData.ts
/**
 * Generated by orval v6.23.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface MasExternalPersonalData {
    /** ชื่อหน่วยงาน */
    depName?: string;
    /** ใช้แสดงใน auto complete  */
    displayText?: string;
    /** อีเมล */
    email?: string;
    /** Filename */
    filename?: string;
    /** ชื่อ (ภาษาอังกฤษ) */
    fullnameEn?: string;
    /** ชื่อ (ภาษาไทย) */
    fullnameTh?: string;
    /** PK */
    personId?: number;
    /** Module */
    module?: number;
    /** Prefix */
    prefix?: string;
    /** ค่าที่ใช้ค้นหา */
    searchText?: string;
}